feat(bin-designer): expose text relief depth in the cutout editor
The cutout editor's label controls offered engrave/emboss with no way to
set how deep the cut or how tall the relief is; the only knob lived in the
main panel's Typography disclosure. Surface the design-wide depth next to
the mode toggle, captioned for the active mode (engrave depth / emboss
height).

Label controls move to their own file so the inspector stays under the
line cap.
