+++ src/i18n/locales/cs.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Vlevo nahoře",
   "binDesigner.cutoutTextAnchor.top-right": "Vpravo nahoře",
   "binDesigner.cutoutTextAngle": "Úhel",
+  "binDesigner.cutoutTextDepth.emboss": "Výška vyvýšení",
+  "binDesigner.cutoutTextDepth.engrave": "Hloubka gravírování",
+  "binDesigner.cutoutTextDepth.hint": "Platí pro veškerý text v tomto návrhu",
   "binDesigner.cutoutTextOffsetX": "Posun X",
   "binDesigner.cutoutTextOffsetY": "Posun Y",
   "binDesigner.cutouts.acrossFlats": "Přes plošky",
