+++ src/i18n/locales/de.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Oben links",
   "binDesigner.cutoutTextAnchor.top-right": "Oben rechts",
   "binDesigner.cutoutTextAngle": "Winkel",
+  "binDesigner.cutoutTextDepth.emboss": "Prägehöhe",
+  "binDesigner.cutoutTextDepth.engrave": "Gravurtiefe",
+  "binDesigner.cutoutTextDepth.hint": "Gilt für den gesamten Text in diesem Design",
   "binDesigner.cutoutTextOffsetX": "Versatz X",
   "binDesigner.cutoutTextOffsetY": "Versatz Y",
   "binDesigner.cutouts.acrossFlats": "Schlüsselweite",
