+++ src/i18n/locales/en.ts
@@ -1942,6 +1942,9 @@ const en: Record<string, string> = {
   'binDesigner.cutoutTextOffsetX': 'Nudge X',
   'binDesigner.cutoutTextOffsetY': 'Nudge Y',
   'binDesigner.cutoutTextAngle': 'Angle',
+  'binDesigner.cutoutTextDepth.engrave': 'Engrave depth',
+  'binDesigner.cutoutTextDepth.emboss': 'Emboss height',
+  'binDesigner.cutoutTextDepth.hint': 'Applies to all text in this design',
   'binDesigner.cutoutLabelMode': 'Label type',
   'binDesigner.cutoutLabelMode.engrave': 'Engrave',
   'binDesigner.cutoutLabelMode.socket': 'Socket',
