+++ src/i18n/locales/es.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Arriba izq.",
   "binDesigner.cutoutTextAnchor.top-right": "Arriba der.",
   "binDesigner.cutoutTextAngle": "Ángulo",
+  "binDesigner.cutoutTextDepth.emboss": "Altura del relieve",
+  "binDesigner.cutoutTextDepth.engrave": "Profundidad de grabado",
+  "binDesigner.cutoutTextDepth.hint": "Se aplica a todo el texto de este diseño",
   "binDesigner.cutoutTextOffsetX": "Desplazar X",
   "binDesigner.cutoutTextOffsetY": "Desplazar Y",
   "binDesigner.cutouts.acrossFlats": "Entrecaras",
