+++ src/i18n/locales/fr.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "En haut à gauche",
   "binDesigner.cutoutTextAnchor.top-right": "En haut à droite",
   "binDesigner.cutoutTextAngle": "Angle",
+  "binDesigner.cutoutTextDepth.emboss": "Hauteur du relief",
+  "binDesigner.cutoutTextDepth.engrave": "Profondeur de gravure",
+  "binDesigner.cutoutTextDepth.hint": "S'applique à tout le texte de ce design",
   "binDesigner.cutoutTextOffsetX": "Décalage X",
   "binDesigner.cutoutTextOffsetY": "Décalage Y",
   "binDesigner.cutouts.acrossFlats": "Entreplats",
