+++ src/i18n/locales/it.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "In alto a sinistra",
   "binDesigner.cutoutTextAnchor.top-right": "In alto a destra",
   "binDesigner.cutoutTextAngle": "Angolo",
+  "binDesigner.cutoutTextDepth.emboss": "Altezza del rilievo",
+  "binDesigner.cutoutTextDepth.engrave": "Profondità di incisione",
+  "binDesigner.cutoutTextDepth.hint": "Si applica a tutto il testo di questo design",
   "binDesigner.cutoutTextOffsetX": "Sposta X",
   "binDesigner.cutoutTextOffsetY": "Sposta Y",
   "binDesigner.cutouts.acrossFlats": "Tra i piani",
