+++ src/i18n/locales/ja.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "左上",
   "binDesigner.cutoutTextAnchor.top-right": "右上",
   "binDesigner.cutoutTextAngle": "角度",
+  "binDesigner.cutoutTextDepth.emboss": "凸彫りの高さ",
+  "binDesigner.cutoutTextDepth.engrave": "彫刻の深さ",
+  "binDesigner.cutoutTextDepth.hint": "このデザインのすべてのテキストに適用されます",
   "binDesigner.cutoutTextOffsetX": "オフセット X",
   "binDesigner.cutoutTextOffsetY": "オフセット Y",
   "binDesigner.cutouts.acrossFlats": "二面幅",
