+++ src/i18n/locales/ko.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "왼쪽 위",
   "binDesigner.cutoutTextAnchor.top-right": "오른쪽 위",
   "binDesigner.cutoutTextAngle": "각도",
+  "binDesigner.cutoutTextDepth.emboss": "양각 높이",
+  "binDesigner.cutoutTextDepth.engrave": "음각 깊이",
+  "binDesigner.cutoutTextDepth.hint": "이 디자인의 모든 텍스트에 적용됩니다",
   "binDesigner.cutoutTextOffsetX": "X 미세 이동",
   "binDesigner.cutoutTextOffsetY": "Y 미세 이동",
   "binDesigner.cutouts.acrossFlats": "대변 거리",
