+++ src/i18n/locales/nb.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Øverst venstre",
   "binDesigner.cutoutTextAnchor.top-right": "Øverst høyre",
   "binDesigner.cutoutTextAngle": "Vinkel",
+  "binDesigner.cutoutTextDepth.emboss": "Preggehøyde",
+  "binDesigner.cutoutTextDepth.engrave": "Graveringsdybde",
+  "binDesigner.cutoutTextDepth.hint": "Gjelder all tekst i dette designet",
   "binDesigner.cutoutTextOffsetX": "Forskyv X",
   "binDesigner.cutoutTextOffsetY": "Forskyv Y",
   "binDesigner.cutouts.acrossFlats": "Nøkkelvidde",
