+++ src/i18n/locales/nl.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Linksboven",
   "binDesigner.cutoutTextAnchor.top-right": "Rechtsboven",
   "binDesigner.cutoutTextAngle": "Hoek",
+  "binDesigner.cutoutTextDepth.emboss": "Reliëfhoogte",
+  "binDesigner.cutoutTextDepth.engrave": "Graveerdiepte",
+  "binDesigner.cutoutTextDepth.hint": "Geldt voor alle tekst in dit ontwerp",
   "binDesigner.cutoutTextOffsetX": "Verschuif X",
   "binDesigner.cutoutTextOffsetY": "Verschuif Y",
   "binDesigner.cutouts.acrossFlats": "Sleutelwijdte",
