+++ src/i18n/locales/pl.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Lewy górny",
   "binDesigner.cutoutTextAnchor.top-right": "Prawy górny",
   "binDesigner.cutoutTextAngle": "Kąt",
+  "binDesigner.cutoutTextDepth.emboss": "Wysokość wypukłości",
+  "binDesigner.cutoutTextDepth.engrave": "Głębokość graweru",
+  "binDesigner.cutoutTextDepth.hint": "Dotyczy całego tekstu w tym projekcie",
   "binDesigner.cutoutTextOffsetX": "Dosuń X",
   "binDesigner.cutoutTextOffsetY": "Dosuń Y",
   "binDesigner.cutouts.acrossFlats": "Między płaskami",
