+++ src/i18n/locales/pt-BR.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Sup. esquerda",
   "binDesigner.cutoutTextAnchor.top-right": "Sup. direita",
   "binDesigner.cutoutTextAngle": "Ângulo",
+  "binDesigner.cutoutTextDepth.emboss": "Altura do relevo",
+  "binDesigner.cutoutTextDepth.engrave": "Profundidade de gravação",
+  "binDesigner.cutoutTextDepth.hint": "Aplica-se a todo o texto deste design",
   "binDesigner.cutoutTextOffsetX": "Deslocar X",
   "binDesigner.cutoutTextOffsetY": "Deslocar Y",
   "binDesigner.cutouts.acrossFlats": "Entre planos",
