+++ src/i18n/locales/sv.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Uppe till vänster",
   "binDesigner.cutoutTextAnchor.top-right": "Uppe till höger",
   "binDesigner.cutoutTextAngle": "Vinkel",
+  "binDesigner.cutoutTextDepth.emboss": "Präglingshöjd",
+  "binDesigner.cutoutTextDepth.engrave": "Graveringsdjup",
+  "binDesigner.cutoutTextDepth.hint": "Gäller all text i denna design",
   "binDesigner.cutoutTextOffsetX": "Förskjut X",
   "binDesigner.cutoutTextOffsetY": "Förskjut Y",
   "binDesigner.cutouts.acrossFlats": "Nyckelvidd",
