+++ src/i18n/locales/uk.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "Угорі ліворуч",
   "binDesigner.cutoutTextAnchor.top-right": "Угорі праворуч",
   "binDesigner.cutoutTextAngle": "Кут",
+  "binDesigner.cutoutTextDepth.emboss": "Висота тиснення",
+  "binDesigner.cutoutTextDepth.engrave": "Глибина гравіювання",
+  "binDesigner.cutoutTextDepth.hint": "Застосовується до всього тексту в цьому дизайні",
   "binDesigner.cutoutTextOffsetX": "Зсув X",
   "binDesigner.cutoutTextOffsetY": "Зсув Y",
   "binDesigner.cutouts.acrossFlats": "Розмір під ключ",
