+++ src/i18n/locales/zh-CN.json
@@ -636,6 +636,9 @@
   "binDesigner.cutoutTextAnchor.top-left": "左上",
   "binDesigner.cutoutTextAnchor.top-right": "右上",
   "binDesigner.cutoutTextAngle": "角度",
+  "binDesigner.cutoutTextDepth.emboss": "浮雕高度",
+  "binDesigner.cutoutTextDepth.engrave": "雕刻深度",
+  "binDesigner.cutoutTextDepth.hint": "适用于此设计中的所有文本",
   "binDesigner.cutoutTextOffsetX": "微调 X",
   "binDesigner.cutoutTextOffsetY": "微调 Y",
   "binDesigner.cutouts.acrossFlats": "对边宽度",
